chatbot-agent/pyproject.toml
[project]
name = "basic-agent-chainlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.124.0",
    "openai-agents>=0.0.19",
    "python-dotenv>=1.1.0",
    "qdrant-client>=1.16.1",
    "asyncpg>=0.29.0",
    "bcrypt>=4.0.1",
    "cryptography>=41.0.0",
]
